move get_index_value to vyper/context/definitions/utils
